Adapted Echo search and DOA

+++ src/hardtarget/target_estimation/dpt/dpt_process.py
@@ -135,8 +135,8 @@ class DPTProcess(TargetEstimationProcess[DPTCfgParams, DPTProParams]):
             return MFVariables(
                 vals=np.zeros((len(self.pro_params.ranges),), dtype=np.float32),
                 dc=np.zeros((len(self.pro_params.ranges),), dtype=np.float32),
-                v=np.zeros((len(self.pro_params.ranges),), dtype=np.float32),
-                a=np.zeros((len(self.pro_params.ranges),), dtype=np.float32),
-                phi=np.zeros((len(self.pro_params.ranges),), dtype=np.float32),
+                v=np.zeros((len(self.pro_params.ranges),), dtype=np.float64),
+                a=np.zeros((len(self.pro_params.ranges),), dtype=np.float64),
+                phi=np.zeros((len(self.pro_params.ranges),), dtype=np.float64),
                 tx_pwr=np.array(tx_pwr),
             )
